[기능개선][Entity] Entity -> Dto 변환 시 MapStruct 사용 feat: MapStruct 도입 및 공연, 공연장 기존 변환 로직 수정 #78
