+++ build.gradle
@@ -67,6 +67,13 @@ dependencies {
 
     // Validation
     implementation 'org.springframework.boot:spring-boot-starter-validation'
+
+    // MapStruct 라이브러리
+    implementation 'org.mapstruct:mapstruct:1.5.5.Final'
+
+    // MapStruct 어노테이션 프로세서
+    annotationProcessor 'org.mapstruct:mapstruct-processor:1.5.5.Final'
+
 }
 
 tasks.named('test') {
